app/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.9
django-allauth==0.41.0
django-celery==3.3.1
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-excel==0.0.10
django-import-export==1.2.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
kombu==3.0.37
lml==0.0.9
MarkupPy==1.14
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
psycopg2==2.8.4
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
tablib==0.14.0
texttable==1.6.2
urllib3==1.25.7
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
